% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_create.R
\name{tl_create}
\alias{tl_create}
\title{Create a Lightning Memory-Mapped Database (LMDB) for a given provider}
\usage{
tl_create(
  provider = getOption("tl_default_provider", "itis"),
  version = tl_latest_version(),
  dir = tl_dir(),
  lines = 500000L
)
}
\arguments{
\item{provider}{Abbreviation for a known naming provider.
Provider data should first be imported with \verb{[tl_create]}.
Note: setting provider to \code{"itis_test"} is for testing purposes only,
use \code{"itis"} for the full ITIS data.  See details}

\item{version}{version of the authority to use (e.g. four-digit year)}

\item{dir}{storage location for the LMDB databases}
}
\description{
Download raw data and store in a local LMDB database.  Importing
data is a time-consuming step that needs be run only once per
machine and will persist through sessions.
}
\examples{

## example uses "itis_test" for illustration only:
tl_create("itis_test")


}
