% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tl_dir.R
\name{tl_dir}
\alias{tl_dir}
\title{taxalight data directory}
\usage{
tl_dir()
}
\description{
taxalight stores data for persistent access in the directory given by
tl_dir() by default.  All functions can override this choice by passing
an alternative path to the \code{dir} argument, or configure the location
system-wide by setting the environmental variable \code{TAXALIGHT_HOME},
eg. in their \code{.Renviron} file, see \code{\link[=Sys.setenv]{Sys.setenv()}}.  If unset, the default
location is the default for the operating system, as provided by the
core R function \code{\link[tools:userdir]{tools::R_user_dir()}}.  Users can manually purge the data
storage at any time by deleting this directory.
}
\examples{
tl_dir()

}
